Percona Server release 5.7.23-24

1. git merge Percona-Server-5.7.23-24

The detail merge info:

```
warning: Failed to merge submodule storage/tokudb/PerconaFT (not fast-forward)
Found a possible merge resolution for the submodule:
 44825986913813b0995bae41f93d6f5bef5787e4: > Merge remote-tracking branch 'percona/master'
If this is correct simply add it to the index for example
by using:

  git update-index --cacheinfo 160000 44825986913813b0995bae41f93d6f5bef5787e4 "storage/tokudb/PerconaFT"

which will accept this suggestion.
自动合并 storage/tokudb/tokudb_sysvars.h
自动合并 storage/tokudb/tokudb_sysvars.cc
自动合并 storage/tokudb/hatoku_hton.h
自动合并 storage/tokudb/hatoku_hton.cc
冲突（内容）：合并冲突于 storage/tokudb/hatoku_hton.cc
删除 storage/tokudb/ha_tokudb_mrr_maria.cc
冲突（修改/删除）：storage/tokudb/ha_tokudb_alter_common.cc 在 Percona-Server-5.7.23-24 中被删除，在 HEAD 中被 修改。storage/tokudb/ha_tokudb_alter_common.cc 在 HEAD 中的版本被保留。
删除 storage/tokudb/ha_tokudb_alter_55.cc
自动合并 storage/tokudb/ha_tokudb_alter.cc
自动合并 storage/tokudb/ha_tokudb.cc
冲突（内容）：合并冲突于 storage/tokudb/ha_tokudb.cc
自动合并 storage/tokudb/PerconaFT
冲突（子模组）：合并冲突于 storage/tokudb/PerconaFT
自动合并 storage/tokudb/CMakeLists.txt
自动合并 storage/myisam/sort.cc
自动合并 sql/sql_yacc.yy
自动合并 sql/sql_show.cc
自动合并 sql/mysqld.cc
自动合并 sql/lex.h
自动合并 sql/handler.h
自动合并 sql/handler.cc
自动合并 sql-common/get_password.c
删除 rapid/plugin/x/mysqlxtest_src/mysqlx_common.h
删除 plugin/keyring_vault/vault_secure_string.h
删除 plugin/keyring_vault/vault_memory.h
删除 plugin/keyring_vault/tests/mtr/vault_cleanup.inc
删除 plugin/keyring/keyring-test/keyring-test.cc
自动合并 packaging/rpm-docker/my.cnf.in
自动合并 packaging/WiX/configure_file.cmake
删除 packaging/WiX/ca/CustomAction.def
删除 packaging/WiX/ca/CustomAction.cpp
删除 packaging/WiX/ca/CMakeLists.txt
删除 mysql-test/t/percona_log_slow_innodb.test
删除 mysql-test/t/bug86260.test
删除 mysql-test/suite/tokudb/t/fast_upsert_int.py
删除 mysql-test/suite/tokudb/t/fast_update_varchar.py
删除 mysql-test/suite/tokudb/t/fast_update_int.py
删除 mysql-test/suite/tokudb/t/fast_update_disable_slow_update.test
删除 mysql-test/suite/tokudb/t/fast_update_decr_floor.py
删除 mysql-test/suite/tokudb/t/fast_update_blobs_with_varchar.py
删除 mysql-test/suite/tokudb/t/fast_update_blobs_fixed_varchar.test
删除 mysql-test/suite/tokudb/t/fast_update_blobs_fixed_varchar.py
删除 mysql-test/suite/tokudb/t/fast_update_blobs.py
删除 mysql-test/suite/tokudb/r/fast_update_disable_slow_update.result
删除 mysql-test/suite/tokudb/r/fast_update_blobs_fixed_varchar.result
自动合并 mysql-test/suite/tokudb.rpl/t/rpl_xa_interleave.test
冲突（添加/添加）：合并冲突于 mysql-test/suite/tokudb.rpl/t/rpl_xa_interleave.test
自动合并 mysql-test/suite/tokudb.rpl/t/rpl_stmt_replace_into.test
自动合并 mysql-test/suite/tokudb.rpl/t/rpl_row_replace_into.test
自动合并 mysql-test/suite/tokudb.rpl/t/rpl_mixed_replace_into.test
自动合并 mysql-test/suite/tokudb.rpl/r/rpl_xa_interleave.result
冲突（添加/添加）：合并冲突于 mysql-test/suite/tokudb.rpl/r/rpl_xa_interleave.result
自动合并 mysql-test/suite/tokudb.rpl/r/rpl_stmt_replace_into.result
自动合并 mysql-test/suite/tokudb.rpl/r/rpl_row_replace_into.result
自动合并 mysql-test/suite/tokudb.rpl/r/rpl_mixed_replace_into.result
自动合并 mysql-test/suite/sys_vars/r/all_vars.result
自动合并 mysql-test/suite/rpl/t/percona_log_slow_slave_statements_innodb.test
删除 mysql-test/suite/rpl/r/bug84415_logical_clock.result
删除 mysql-test/suite/rpl/r/bug84415.result
删除 mysql-test/suite/rocksdb/t/insert_optimized_config-master.opt
删除 mysql-test/suite/rocksdb/t/autoincrement.test
删除 mysql-test/suite/rocksdb/r/autoincrement.result
删除 mysql-test/suite/rocksdb.sys_vars/t/rocksdb_flush_memtable_on_analyze_basic.test
删除 mysql-test/suite/rocksdb.sys_vars/r/rocksdb_flush_memtable_on_analyze_basic.result
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_barracuda_redundant_set3.test
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_barracuda_dynamic_set3.test
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_barracuda_compressed_set3.test
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_barracuda_compact_set3.test
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_antelope_redundant_set3.test
自动合并 mysql-test/suite/innodb/t/xtradb_compressed_columns_consistency_antelope_compact_set3.test
删除 mysql-test/suite/innodb/t/innodb-wl6045-1.test
删除 mysql-test/suite/innodb/t/innodb-wl6045-1-master.opt
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_barracuda_compressed_set1.result
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_barracuda_compact_set3.result
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_barracuda_compact_set2.result
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_barracuda_compact_set1.result
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_antelope_compact_set2.result
自动合并 mysql-test/suite/innodb/r/xtradb_compressed_columns_consistency_antelope_compact_set1.result
删除 mysql-test/suite/innodb/r/percona_innodb_kill_idle_trx.result
自动合并 mysql-test/suite/audit_null/t/audit_plugin_bugs.test
自动合并 mysql-test/suite/audit_null/t/audit_plugin_2.test
自动合并 mysql-test/suite/audit_null/r/audit_plugin_bugs.result
自动合并 mysql-test/suite/audit_null/r/audit_plugin_2.result
删除 mysql-test/r/percona_log_slow_slave_statements-innodb.result
删除 mysql-test/r/percona_log_slow_innodb.result
删除 mysql-test/r/bug86260.result
自动合并 mysql-test/lib/mtr_cases.pm
删除 mysql-test/include/check_openssl.inc
自动合并 mysql-test/extra/rpl_tests/rpl_seconds_behind_master_mts_type_database.test
自动合并 mysql-test/extra/rpl_tests/rpl_seconds_behind_master_mts_logical_clock.test
删除 man/mysqltest_embedded.1
删除 man/mysqltest.1
删除 man/mysql_client_test_embedded.1
删除 man/mysql_client_test.1
删除 man/mysql-test-run.pl.1
删除 man/mysql-stress-test.pl.1
删除 cmake/merge_archives_unix.cmake.in
自动合并 cmake/maintainer.cmake
自动合并 cmake/build_configurations/compiler_options.cmake
冲突（内容）：合并冲突于 cmake/build_configurations/compiler_options.cmake
自动合并 build-ps/percona-server.spec
冲突（内容）：合并冲突于 build-ps/percona-server.spec
自动合并 build-ps/build-binary.sh
自动合并 README.md
冲突（内容）：合并冲突于 README.md
自动合并 CMakeLists.txt
自动合并 .travis.yml
冲突（添加/添加）：合并冲突于 .travis.yml
自动合并失败，修正冲突然后提交修正的结果。
```

+++ rapid/plugin/group_replication/src/applier.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014, 2017, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2014, 2018, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -298,6 +298,11 @@ int Applier_module::apply_data_packet(Data_packet *data_packet,
   int error= 0;
   uchar* payload= data_packet->payload;
   uchar* payload_end= data_packet->payload + data_packet->len;
+
+  DBUG_EXECUTE_IF("group_replication_before_apply_data_packet", {
+    const char act[] = "now wait_for continue_apply";
+    DBUG_ASSERT(!debug_sync_set_action(current_thd, STRING_WITH_LEN(act)));
+  });
 
   if (check_single_primary_queue_status())
     return 1; /* purecov: inspected */
@@ -786,7 +791,7 @@ Applier_module::wait_for_applier_complete_suspension(bool *abort_flag,
   {
     error= APPLIER_GTID_CHECK_TIMEOUT_ERROR; //timeout error
     while (error == APPLIER_GTID_CHECK_TIMEOUT_ERROR && !(*abort_flag))
-      error= wait_for_applier_event_execution(1); //blocking
+      error= wait_for_applier_event_execution(1, true); //blocking
   }
 
   return (error == APPLIER_RELAY_LOG_NOT_INITED);
@@ -816,7 +821,8 @@ Applier_module::is_applier_thread_waiting()
 }
 
 int
-Applier_module::wait_for_applier_event_execution(double timeout)
+Applier_module::wait_for_applier_event_execution(double timeout,
+                                                 bool check_and_purge_partial_transactions)
 {
   DBUG_ENTER("Applier_module::wait_for_applier_event_execution");
   int error= 0;
@@ -834,7 +840,8 @@ Applier_module::wait_for_applier_event_execution(double timeout)
       the applier thread will release the lock and update the applier thread
       execution position correctly and safely.
     */
-    if (((Applier_handler*)event_applier)->is_partial_transaction_on_relay_log())
+    if (check_and_purge_partial_transactions &&
+        ((Applier_handler*)event_applier)->is_partial_transaction_on_relay_log())
     {
         error= purge_applier_queue_and_restart_applier_module();
     }
